refactor: extract OrganizationsPage into new component
