+++ src/pages/admin/organizations/index.tsx
@@ -1,12 +1,7 @@
-import apiClient from "@api/client";
-import { Organization } from "@api/client/models/Organization";
-import OrganizationTable from "@components/OrganizationTable";
-import Page from "@components/Page";
+import OrganizationsPage from "@components/OrganizationsPage";
 import { loadMessages } from "@utils/i18n";
 import withAuth from "@utils/withAuth";
 import { GetStaticProps } from "next";
-import { useEffect, useState } from "react";
-import { useTranslations } from "use-intl";
 
 export const getStaticProps: GetStaticProps = async (context) => ({
 	props: {
@@ -14,30 +9,4 @@ export const getStaticProps: GetStaticProps = async (context) => ({
 	},
 });
 
-const OrganizationList = () => {
-	const t = useTranslations("Organizations");
-	const [organizations, setOrganizations] = useState<Organization[] | undefined>(undefined);
-
-	const fetchOrganizations = () => {
-		apiClient.discoverCulturalData.getOrganizations().then((res) => {
-			const organizations = res?.data?.organizations;
-			setOrganizations(organizations);
-		});
-	};
-
-	useEffect(() => {
-		fetchOrganizations();
-	}, []);
-
-	return (
-		<Page metadata={{ title: t("page-title") }}>
-			{organizations ? (
-				<OrganizationTable organizations={organizations} fetchOrganizations={fetchOrganizations} />
-			) : (
-				<div>Loading...</div>
-			)}
-		</Page>
-	);
-};
-
-export default withAuth(OrganizationList);
+export default withAuth(OrganizationsPage);
